pyproject.toml
[tool.poetry]
name = "typeright"
version = "0.1.0"
description = "Add type annotations to your source code"
authors = ["Chad Dombrova <chadrik@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/chadrik/typeright"
keywords = ["pep484", "typing", "annotations", "mypy"]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Environment :: Console',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: POSIX',
    'Programming Language :: Python :: 2',
    'Programming Language :: Python :: 2.7',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Topic :: Software Development',
]

[tool.poetry.scripts]
typeright = 'typeright.__main__:main'

[tool.poetry.dependencies]
python = "~2.7 || ^3.6"
six = "*"
docutils = "*"
"kids.cache" = "*"
typing = { version = "3.7.*", python = "<3.7" }

pytest = { version = "^6.1", optional = true }
coverage = { version = "*", optional = true }
pytest-mock = { version = "*", optional = true }
mypy_extensions = { version = ">=0.3.0", optional = true }
typing-extensions = { version = ">=3.7.4.3", optional = true }

[tool.poetry.extras]
tests = [
    "pytest",
    "coverage",
    "pytest-mock",
    "mypy_extensions",
    "typing-extensions"
]

[tool.poetry.dev-dependencies]
tox = "^3.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
